+++ cinecore/modelos/IngressoMeia.cs
@@ -8,8 +8,8 @@ namespace cinecore.modelos
     public class IngressoMeia : Ingresso
     {
         [Required(ErrorMessage = "O preço é obrigatório")]
-        [Range(0, float.MaxValue, ErrorMessage = "O preço não pode ser negativo")]
-        public float Preco { get; set; }
+        [Range(0, double.MaxValue, ErrorMessage = "O preço não pode ser negativo")]
+        public decimal Preco { get; set; }
 
         [Required(ErrorMessage = "O motivo é obrigatório")]
         [StringLength(200, MinimumLength = 1, ErrorMessage = "O motivo deve ter entre 1 e 200 caracteres")]
@@ -17,7 +17,7 @@ namespace cinecore.modelos
 
         public IngressoMeia() : base() { }
 
-        public IngressoMeia(float preco, int id, char fila, int numero, DateTime dataCompra, string motivo,
+        public IngressoMeia(decimal preco, int id, char fila, int numero, DateTime dataCompra, string motivo,
             Sessao? sessao = null, Cliente? cliente = null, Assento? assento = null)
             : base(id, fila, numero, dataCompra, sessao, cliente, assento)
         {
@@ -25,7 +25,7 @@ namespace cinecore.modelos
             Motivo = motivo;
         }
 
-        public override float CalcularPreco(float precoBase)
+        public override decimal CalcularPreco(decimal precoBase)
         {
             return Preco / 2;
         }
